Connection opens only when used
Over the years I have kept the connection open. Now,
everytime need to communicate with the device, I open the
connection, do the transaction and close the socket.
Apparently this is the right thing to do. Opening and closing
the sockets is very cheap. Things appear to be working.
I will try to clean up the code, document things. I have
no Idea what kind of tests I can do...