Corrected nasty bug: OrderedDict vs Dict
When calling daqaddinput, the method was using a Dict and then it lost
track of ordering. When waving data with savedaqdata it changed
the order of the channels so that when reading back the data, it
was in the wrong order.

I really should implement a DAQChannels data structure to handle this all.